backend/requirements.txt
Flask==2.2.2
Flask_Cors==3.0.10
PyJWT==2.6.0
pymysql==1.0.2